requirements.txt
openai_whisper==20250625
streamlit==1.51.0
imageio-ffmpeg
